deepdream-generator-api/deepdream_generator_api/functions/deepdream/requirements.txt
pynamodb
tensorflow
Pillow
requests_aws4auth
gql
